Changed classes design for clearer responsibilities
Two major design changes were made:
 1. The Person model is no longer responsible for validating its ID
 2. PersonRepository is now an abstract class providing base behavior

The Person model was internally parsing its ID as a GUID in the
constructor, and the other classes were using Guid instances throughout.
The Guid usages were changed to use simple string, and only the Guid
class is only used for validation, which is now a responsibility of a
separate class, PersonValidator, to which the PersonRepository delegates
validation of new Person instances.

The PersonRepository class is now abstract, and does not weigh in on the
actual storing behavior, only taking care of validation and making sure
that there are no duplicate items in the repository (Person instances
are compared for equality by their GUID) and that basic CRUD operation
rules are respected in the context of Person equality criteria (a Person
can only be updated in the repository if its ID hasn't changed, in which
case it'd be a distinct person).

Abstracting the storage method away from the business rules of the
PersonRepository was actually motivated by the fact that in the future
the system will be changed to persist the data.