Redesigned the ICommand interface
The ICommand interface was redesigned to handle its own action, thus
solving the previous design's issue in which command arguments were
impossible (each command would need different arguments, so declaring
them in the ICommand interface would not be feasible). Now Commands
encapsulate their own logic and execute themselves (when told to do so).

To enable commands to handle themselves, they must interact directly
with the user. Dependencies such as a Repository and a TextReader are
injected through Property Injection.