Added a Dependencies property to ICommand
A Dependencies property was added to ICommand. This property is an
IEnumerable of ICommands, intended to be a way for commands to declare
actions that need to precede their own.

The first use case will be having the new DeletePerson command declare
ListPeople as a dependency. All the people in the Repository will be
listed before the user is asked which person he'd like to delete.