Fixed broken Repository tests
-1. Fixed a bug in the Validator mock:
Apparently, using It.IsAny<> through a variable didn't setup the mock
correctly.

-2. Fixed a bug in the PersonRepository abstract class:
For some reason, the default IEqualityComparer<Person> in the Repository
HashSet was not being called for the Update operation. As a quick fix,
the Update operation will be validated explicitly by the Id property
of the two Person instances, instead of using Linq's Contains() method.

Other changes:
 - Renamed the Id property of the Person class, to match the naming
 convention.