Fixed a bug in DeletePerson command
The ordering in ListPerson and DeletePerson command were different.
Since these two commands are currently coupled, the user was being
misled and was deleting the wrong birthday. The user was displayed a
list in ascending order and then asked for the index of the birthday to
be deleted, but then in the DeletePerson command the index would be read
of a list in inverse order, hence almost always DeletePerson removed the
wrong entry. This was fixed by making the DeletePerson also read from an
ascending order list (expected). Coupling is still an issue though.

Other changes:
 - Changed DeletePerson to show the name of the deleted person.